requirements.txt
tensorflow
streamlit
opencv-python
Pillow
numpy
pandas
matplotlib
seaborn
plotly
scikit-learn
pytest
jupyter